@@ -10,19 +10,19 @@
     "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^3.0.2",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tsconfig/svelte": "^5.0.2",
     "autoprefixer": "^10.4.19",
     "colorjs.io": "^0.5.0",
     "daisyui": "^4.11.1",
-    "svelte": "^4.2.12",
-    "svelte-check": "^3.6.7",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "svelte-legos": "^0.2.2",
     "svelte-lucide": "^1.0.5",
     "tailwindcss": "^3.4.3",
     "tslib": "^2.6.2",
-    "typescript": "^5.2.2",
-    "vite": "^5.2.0",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4",
     "vite-plugin-singlefile": "^2.0.1"
   }
 }
